Fixing warnings: mismatch integer/pointer size
The thandle_t type is defined (in some platforms) as void *,
and this type is converted to (int) in several places in the
code. This conversion is not valid in 64 bits machines.
We have introduced intermediate castings, but they simply
silence the warning. The underlying risk of truncating
(and loosing information) in the casting process is still there.

The choice of "void *" as the type for thandle_t is questionable...
particularly given that it is commonly used as a file descriptor.